requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
lightgbm==4.5.0
numexpr==2.10.1
matplotlib==3.8.4

pillow==10.4.0
onnxruntime==1.19.2
requests==2.32.3

# Testing
pytest==8.0.0
httpx==0.27.0
